feat: subspaces gallery
