+++ apps/web/partials/space-page/subspaces.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import { SYSTEM_IDS } from '@geogenesis/ids';
+import Image from 'next/image';
 import Link from 'next/link';
 
 import { useSpaces } from '~/core/hooks/use-spaces';
@@ -16,8 +17,6 @@ type SubspacesProps = {
 export const Subspaces = ({ subspaces }: SubspacesProps) => {
   const { spaces } = useSpaces();
 
-  if (subspaces.length === 0) return null;
-
   return (
     <>
       <Slider label="Subspaces">
@@ -32,9 +31,16 @@ export const Subspaces = ({ subspaces }: SubspacesProps) => {
             space.attributes[SYSTEM_IDS.IMAGE_ATTRIBUTE] ?? 'https://via.placeholder.com/600x600/FF00FF/FFFFFF';
 
           return (
-            <Link key={index} href={href}>
+            <Link key={index} href={href} className="group">
               <div className="relative aspect-[16/9] w-full overflow-clip rounded bg-grey-01">
-                <img src={src} className="absolute inset-0 h-full w-full object-cover object-center" alt="" />
+                <Image
+                  src={src}
+                  className="transition-transform duration-150 ease-in-out group-hover:scale-105"
+                  objectFit="cover"
+                  priority
+                  layout="fill"
+                  alt=""
+                />
               </div>
               <div className="mt-2 text-metadataMedium text-text">{subspace.name}</div>
               {/* <div className="mt-1 flex items-center gap-1 text-smallButton text-grey-03">
